LayoutTests/imported/w3c/web-platform-tests/intersection-observer/zero-area-element-hidden-expected.txt


PASS A zero-area hidden target should not be intersecting. 
FAIL First rAF. assert_equals: entries[0].boundingClientRect.left expected 8 but got 0
